Fixed namespaces and moved some files around. More to come
