+++ ReadingList/Helpers/IReadingListCache.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using ReadingList.Models;
 
 namespace ReadingList.Helpers
 {
