+++ ReadingList/Helpers/ReadingListCache.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using ReadingList.Models;
 
 namespace ReadingList.Helpers
 {
