+++ ReadingList/Models/Book.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace ReadingList
+namespace ReadingList.Models
 {
 	public class Book
 	{
