+++ ReadingList/Models/ReadingBoard.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using ReadingList.Models;
 
 namespace ReadingList
 {
