+++ ReadingList/ReadingList.csproj
@@ -176,6 +176,9 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Helpers\ReadingListServiceProfiler.cs" />
+    <Compile Include="Models\IBookFactory.cs" />
+    <Compile Include="Modules\IndexModel.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CompositionRoot.cs" />
@@ -189,7 +192,6 @@
     <Compile Include="Trello\TrelloBoardConstans.cs" />
     <Compile Include="Modules\ModuleExtensions.cs" />
     <Compile Include="Models\Book.cs" />
-    <Compile Include="Helpers\IBookFactory.cs" />
     <Compile Include="Trello\BookFactory.cs" />
     <Compile Include="Trello\ITrelloWebHookSources.cs" />
     <Compile Include="Trello\TrelloWebHookSourcesConfigFileReader.cs" />
@@ -202,12 +204,10 @@
     <Compile Include="Trello\WebHookCaller.cs" />
     <Compile Include="Trello\ITrelloAuthModel.cs" />
     <Compile Include="Trello\TrelloAuthModel.cs" />
-    <Compile Include="Trello\ReadingListServiceProfiler.cs" />
     <Compile Include="Trello\CachedReadingListService.cs" />
     <Compile Include="Models\ReadingBoard.cs" />
     <Compile Include="Trello\IReadingBoardService.cs" />
     <Compile Include="Trello\ReadingBoardService.cs" />
-    <Compile Include="Models\IndexModel.cs" />
     <Compile Include="Modules\RouteModule.cs" />
     <Compile Include="Modules\CallbackModule.cs" />
     <Compile Include="Helpers\IReadingListCache.cs" />
