+++ ReadingList/Trello/ITrelloAuthModel.cs
@@ -1,4 +1,4 @@
-namespace ReadingList
+namespace ReadingList.Trello
 {
 	/// <summary>
 	/// Trello auth model.
