+++ ReadingList/Trello/ITrelloAuthorizationWrapper.cs
@@ -1,3 +1,5 @@
+using ReadingList.Trello;
+
 namespace ReadingList
 {
 	/// <summary>
